cleanup buildSrc a bit and do not do network accesses during task configueration
