+++ build.gradle.kts
@@ -9,26 +9,6 @@ buildscript {
 
 ApiCompatibility.apply(rootProject)
 
-abstract class DownloadFileTask : DefaultTask() {
-  @get:Input
-  abstract val url: Property<String>
-
-  @get:org.gradle.api.tasks.OutputFile
-  abstract val output: RegularFileProperty
-
-  @TaskAction
-  fun taskAction() {
-    val client = okhttp3.OkHttpClient()
-    val request = okhttp3.Request.Builder().get().url(url.get()).build()
-
-    client.newCall(request).execute().body!!.byteStream().use { body ->
-      output.asFile.get().outputStream().buffered().use { file ->
-        body.copyTo(file)
-      }
-    }
-  }
-}
-
 subprojects {
   apply {
     from(rootProject.file("gradle/dependencies.gradle"))
