+++ buildSrc/build.gradle.kts
@@ -1,20 +1,9 @@
 plugins {
-  java
-  kotlin("jvm") version "1.3.41"
+  kotlin("jvm") version "1.3.72"
 }
 
-buildscript {
-  project.apply {
-    from(file("../gradle/dependencies.gradle"))
-  }
-
-  repositories {
-    mavenCentral()
-  }
-
-  dependencies {
-    classpath(groovy.util.Eval.x(project, "x.dep.kotlin.plugin"))
-  }
+project.apply {
+  from(file("../gradle/dependencies.gradle"))
 }
 
 repositories {
@@ -25,7 +14,7 @@ repositories {
 }
 
 dependencies {
-  implementation(kotlin("stdlib"))
+  implementation(groovy.util.Eval.x(project, "x.dep.kotlin.stdLib"))
   implementation(groovy.util.Eval.x(project, "x.dep.okHttp.okHttp4"))
   implementation(groovy.util.Eval.x(project, "x.dep.moshi.moshi"))
   implementation(groovy.util.Eval.x(project, "x.dep.android.plugin"))
@@ -34,5 +23,4 @@ dependencies {
   implementation(groovy.util.Eval.x(project, "x.dep.sqldelight.plugin"))
   // this plugin is added to the classpath but never applied, it is only used for the closeAndRelease code
   implementation(groovy.util.Eval.x(project, "x.dep.vanniktechPlugin"))
-
 }
